Handle sensitive information in log
In case of unsuccessful login, do not log username at all. Consider a
situation where a user mistakenly provides password in username field.
Now the password would be logged in the log file and can be accessed, so
it is important to not log username upon unsucessful login or atleast if
username is not valid. Not logging the username in all case would not
help us as it is necessary for debugging. Thus, in case of successful
the username has to be encrypted and only then logged to the log file.
By doing so, the system admin can still use the log entry to debug and
at the same time no adversary can gain access to any usernames.